// app/types/product.ts
// export interface Product {
//   id: string;

//   name: string;

//   brand: string;

//   category: string;

//   vehicle: string;

//   image: string;

//   price: number;

//   stock: number;

//   description: string;
// }


export interface Product {
  id: string;

  name: string;

  brand: string;

  category: string;

  vehicle: string;

  image: string;

  price: number;

  stock: number;

  description: string;
}